% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress_data.R
\name{regress_data}
\alias{regress_data}
\title{regress_data}
\usage{
regress_data(data)
}
\arguments{
\item{data}{(data set we imported with import_data function called comb_data)}
}
\value{
lm for regression analysis
}
\description{
This function will perform regression analysis of top five and bottom five states, use comb_data for data.
}
\examples{
regress_data(comb_data)

}
