% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_data.R
\name{stat_data}
\alias{stat_data}
\title{stat_data}
\usage{
stat_data(data)
}
\arguments{
\item{data}{(data set we imported with import_data function called comb_data)}
}
\value{
t.test for statistical analysis
}
\description{
This function will perform statistical analysis of top five and bottom five states, use comb_data for data.
}
\examples{
stat_data(comb_data)
}
