% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_data.R
\name{sum_data}
\alias{sum_data}
\title{sum_data}
\usage{
sum_data(data)
}
\arguments{
\item{data}{(data set we imported with import_data function called comb_data)}
}
\value{
summary top and bottom five states
}
\description{
This function will summarize the imported covid 19 data, use comb_data for data.
}
\examples{
sum_data(comb_data)
}
