% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_data.R
\name{vis_data}
\alias{vis_data}
\title{vis_data}
\usage{
vis_data(data)
}
\arguments{
\item{data}{(data to graph, use comb_data)}
}
\description{
This function will visualize data and analysis, use comb_data as data
}
\examples{
vis_data(comb_data)

}
